requirements.txt
requests
matplotlib
numpy
opencv-contrib-python
Pillow
scipy
scikit-image
mahotas